# docs/api/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AI 文章生成后端：用户/会话、文章三阶段流水线（SSE 进度）、配图、Agent 日志；开发态 VIP Mock 支付与管理端升降 VIP。",
        "title": "Wood Passage Creator API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/admin/statistics/overview": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-statistics"
                ],
                "summary": "系统概览统计（管理员）",
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/wood-passage-creator_internal_module_statistics.Overview"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "wood-passage-creator_internal_module_statistics.Overview": {
            "type": "object",
            "properties": {
                "activeUserCount": {
                    "description": "本周有创作的用户数",
                    "type": "integer"
                },
                "avgDurationMs": {
                    "type": "integer"
                },
                "monthCount": {
                    "type": "integer"
                },
                "quotaUsed": {
                    "type": "integer"
                },
                "successRate": {
                    "description": "0~100",
                    "type": "number"
                },
                "todayCount": {
                    "type": "integer"
                },
                "totalCount": {
                    "type": "integer"
                },
                "totalUserCount": {
                    "type": "integer"
                },
                "vipUserCount": {
                    "type": "integer"
                },
                "weekCount": {
                    "type": "integer"
                }
            }
        },
...
        "wood-passage-creator_internal_pkg_response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "SessionAuth": {
            "description": "Session cookie authentication. Example: session=your-session-id",
            "type": "apiKey",
            "name": "Cookie",
            "in": "header"
        }
    }
}